Perform some trivial code cleanup. Remove a value and code section that were no longer used. Rename and change the type of a value from 32 to 16 bit to match its pfkey counterpart.
git-svn-id: svn://svn.shrew.net/ike/head@37 1bb156eb-19af-db11-b919-0014c23ffdf2

+++ source/iked/ike.io.pfkey.cpp
@@ -1050,7 +1050,7 @@ long _IKED::pfkey_send_getspi( IDB_POLICY * policy, IDB_PH2 * ph2 )
 					// copy the request id
 					//
 
-					proposal->reqid = getspi.sa2.reqid;
+					proposal->reqid = ( uint16_t ) getspi.sa2.reqid;
 				}
 			}
 		}
@@ -1080,7 +1080,7 @@ long _IKED::pfkey_send_getspi( IDB_POLICY * policy, IDB_PH2 * ph2 )
 					// copy the request id
 					//
 
-					proposal->reqid = getspi.sa2.reqid;
+					proposal->reqid = ( uint16_t ) getspi.sa2.reqid;
 
 					//
 					// copy the spi value
